+++ Assets/OxGKit/Utilities/Scripts/Runtime/Timer/DTUpdate.cs
@@ -10,11 +10,10 @@ namespace OxGKit.Utilities.Timer
         public delegate void DeltaTimeUpdate(float dt);
         public delegate void DeltaTimeFixedUpdate(float dt);
 
-        private bool _firstInitStartTime = false;                   // 是否首次初始啟動時間                         
-        public DateTime? startTime { get; protected set; } = null;  // 啟動時間, 自行由主管理程序記錄
-
         public DeltaTimeUpdate onUpdate = null;
         public DeltaTimeFixedUpdate onFixedUpdate = null;
+
+        private DateTime _createTime;
         public float timeSinceStartup { get; protected set; }       // 自啟動以來的時間
         public float timeAtLastFrame { get; protected set; }        // 記錄最後一幀的時間
         protected float _timeScale = 1f;                            // 時間尺度, 預設 = 1
@@ -28,21 +27,30 @@ namespace OxGKit.Utilities.Timer
                 else this._timeScale = value;
             }
         }
+        protected float _targetFrameRate = FIXED_FRAME;
+        public float targetFrameRate
+        {
+            get { return this._targetFrameRate; }
+            set
+            {
+                if (value < 0f) this._targetFrameRate = 0f;
+                else this._targetFrameRate = value;
+            }
+        }
         public float deltaTime { get; protected set; }
         public float fixedDeltaTime { get; protected set; }
         protected CancellationTokenSource _updateCts = null;
 
         protected const float FIXED_FRAME = 60;                     // 固定幀數 (固定 1 秒刷新 60 次, 毫秒單位 => 1000 ms / 60 = 16 ms, 秒數單位 => 1 s / 60 = 0.016 s)
-        protected const float MAX_TIMESCALE = 10;
+        protected const float MAX_TIMESCALE = 1 << 6;
+
+        public DTUpdate()
+        {
+            this._createTime = DateTime.Now;
+        }
 
         public void StartUpdate()
         {
-            if (!this._firstInitStartTime)
-            {
-                this.startTime = DateTime.Now;
-                this._firstInitStartTime = true;
-            }
-
             this._updateCts = new CancellationTokenSource();
             SetInterval(this._updateCts).Forget();
         }
@@ -57,27 +65,31 @@ namespace OxGKit.Utilities.Timer
 
         protected async UniTask SetInterval(CancellationTokenSource cts)
         {
-            await UniTask.WaitUntil(() => { return (Time.timeScale * this.timeScale) > 0f; }, PlayerLoopTiming.Update, (cts == null) ? default : cts.Token);
+            if (Time.timeScale > 0 && this.targetFrameRate > 0 && this.timeScale > 0)
+            {
+                await UniTask.WaitUntil(() => { return (Time.timeScale * this.timeScale) > 0f; }, PlayerLoopTiming.Update, (cts == null) ? default : cts.Token);
 
-            // 幀數率
-            float multiTimeScale = Time.timeScale * this.timeScale;
-            float frameRate = FIXED_FRAME * multiTimeScale;
-            // 計算 fixedDeltaTime
-            this.fixedDeltaTime = 1 / frameRate;
+                // 幀數率
+                float multiTimeScale = Time.timeScale * this.timeScale;
+                float frameRate = this.targetFrameRate * multiTimeScale;
+                // 計算 fixedDeltaTime
+                this.fixedDeltaTime = 1 / frameRate;
 
-            await UniTask.Delay(TimeSpan.FromSeconds(this.fixedDeltaTime), false, PlayerLoopTiming.Update, (cts == null) ? default : cts.Token);
+                await UniTask.Delay(TimeSpan.FromSeconds(this.fixedDeltaTime), false, PlayerLoopTiming.Update, (cts == null) ? default : cts.Token);
 
-            this.onFixedUpdate?.Invoke(this.fixedDeltaTime);
+                this.onFixedUpdate?.Invoke(this.fixedDeltaTime);
 
-            // 計算 deltaTime
-            this.deltaTime = this.timeSinceStartup - this.timeAtLastFrame;
-            this.timeAtLastFrame = this.timeSinceStartup;
+                // 計算 deltaTime
+                this.deltaTime = this.timeSinceStartup - this.timeAtLastFrame;
+                this.timeAtLastFrame = this.timeSinceStartup;
 
-            // 計算經過的時間, 當前時間 - 最一開始的時間 = 啟動到現在的經過時間 (換算為【秒】)
-            var timeSpan = DateTime.Now.Subtract((DateTime)this.startTime);
-            this.timeSinceStartup = (float)timeSpan.TotalSeconds;
+                // 計算經過的時間, 當前時間 - 最一開始的時間 = 啟動到現在的經過時間 (秒)
+                var timeSpan = DateTime.Now.Subtract(this._createTime);
+                this.timeSinceStartup = (float)timeSpan.TotalSeconds;
 
-            this.onUpdate?.Invoke(this.deltaTime);
+                this.onUpdate?.Invoke(this.deltaTime);
+            }
+            else await UniTask.Yield();
 
             SetInterval(cts).Forget();
         }