+++ Assets/OxGKit/Utilities/Scripts/Runtime/Timer/RTUpdate.cs
@@ -11,7 +11,9 @@ namespace OxGKit.Utilities.Timer
 
         public RealTimeUpdate onUpdate = null;
         public RealTimeFixedUpdate onFixedUpdate = null;
-        public float realtimeSinceStartup { get; protected set; } // 自遊戲啟動以來的時間
+
+        private DateTime _createTime;
+        public float timeSinceStartup { get; protected set; }     // 自啟動以來的時間
         public float timeAtLastFrame { get; protected set; }      // 記錄最後一幀的時間
         protected float _timeScale = 1f;                          // 時間尺度, 預設 = 1
         public float timeScale
@@ -24,17 +26,30 @@ namespace OxGKit.Utilities.Timer
                 else this._timeScale = value;
             }
         }
+        protected float _targetFrameRate = FIXED_FRAME;
+        public float targetFrameRate
+        {
+            get { return this._targetFrameRate; }
+            set
+            {
+                if (value < 0f) this._targetFrameRate = 0f;
+                else this._targetFrameRate = value;
+            }
+        }
         public float deltaTime { get; protected set; }
         public float fixedDeltaTime { get; protected set; }
         protected CancellationTokenSource _updateCts = null;
 
         protected const float FIXED_FRAME = 60;                   // 固定幀數 (固定 1 秒刷新 60 次, 毫秒單位 => 1000 ms / 60 = 16 ms, 秒數單位 => 1 s / 60 = 0.016 s)
-        protected const float MAX_TIMESCALE = 10;
+        protected const float MAX_TIMESCALE = 1 << 6;
+
+        public RTUpdate()
+        {
+            this._createTime = DateTime.Now;
+        }
 
         public void StartUpdate()
         {
-            if (!RealTime.IsInitStartupTime()) return;
-
             this._updateCts = new CancellationTokenSource();
             SetInterval(this._updateCts).Forget();
         }
@@ -49,26 +64,30 @@ namespace OxGKit.Utilities.Timer
 
         protected async UniTask SetInterval(CancellationTokenSource cts)
         {
-            await UniTask.WaitUntil(() => { return this.timeScale > 0f; }, PlayerLoopTiming.Update, (cts == null) ? default : cts.Token);
+            if (this.targetFrameRate > 0 && this.timeScale > 0)
+            {
+                await UniTask.WaitUntil(() => { return this.timeScale > 0f; }, PlayerLoopTiming.Update, (cts == null) ? default : cts.Token);
 
-            // 幀數率
-            float frameRate = FIXED_FRAME * this.timeScale;
-            // 計算 fixedDeltaTime
-            this.fixedDeltaTime = 1 / frameRate;
+                // 幀數率
+                float frameRate = this.targetFrameRate * this.timeScale;
+                // 計算 fixedDeltaTime
+                this.fixedDeltaTime = 1 / frameRate;
 
-            await UniTask.Delay(TimeSpan.FromSeconds(this.fixedDeltaTime), true, PlayerLoopTiming.Update, (cts == null) ? default : cts.Token);
+                await UniTask.Delay(TimeSpan.FromSeconds(this.fixedDeltaTime), true, PlayerLoopTiming.Update, (cts == null) ? default : cts.Token);
 
-            this.onFixedUpdate?.Invoke(this.fixedDeltaTime);
+                this.onFixedUpdate?.Invoke(this.fixedDeltaTime);
 
-            // 計算 deltaTime
-            this.deltaTime = this.realtimeSinceStartup - this.timeAtLastFrame;
-            this.timeAtLastFrame = this.realtimeSinceStartup;
+                // 計算 deltaTime
+                this.deltaTime = this.timeSinceStartup - this.timeAtLastFrame;
+                this.timeAtLastFrame = this.timeSinceStartup;
 
-            // 計算經過的現實時間, 當前時間 - 最一開始的時間 = 遊戲啟動到現在的經過時間 (換算為【秒】)
-            var timeSpan = DateTime.Now.Subtract((DateTime)RealTime.startupTime);
-            this.realtimeSinceStartup = (float)timeSpan.TotalSeconds;
+                // 計算經過的時間, 當前時間 - 最一開始的時間 = 啟動到現在的經過時間 (秒)
+                var timeSpan = DateTime.Now.Subtract(this._createTime);
+                this.timeSinceStartup = (float)timeSpan.TotalSeconds;
 
-            this.onUpdate?.Invoke(this.deltaTime);
+                this.onUpdate?.Invoke(this.deltaTime);
+            }
+            else await UniTask.Yield();
 
             SetInterval(cts).Forget();
         }