+++ Assets/OxGKit/Utilities/Scripts/Runtime/Timer/RealTime.cs
@@ -5,8 +5,8 @@ namespace OxGKit.Utilities.Timer
 {
     public static class RealTime
     {
-        private static bool _firstInitStartupTime = false;                  // 是否首次初始啟動時間                            
-        public static DateTime? startupTime { get; private set; } = null;   // 啟動時間, 建議由主程序調用初始 (將會是遊戲啟動時間)
+        private static bool _firstInitStartupTime = false;            // 是否首次初始啟動時間                            
+        public static DateTime timeSinceStartup { get; private set; } // 啟動時間, 建議由主程序調用初始 (將會是遊戲啟動時間)
 
         /// <summary>
         /// Call by Main Monobehaviour Awake
@@ -16,14 +16,14 @@ namespace OxGKit.Utilities.Timer
         {
             if (_firstInitStartupTime) return;
 
-            startupTime = DateTime.Now;
+            timeSinceStartup = DateTime.Now;
 
             _firstInitStartupTime = true;
         }
 
         public static bool IsInitStartupTime()
         {
-            if (startupTime == null)
+            if (!_firstInitStartupTime)
             {
                 Debug.Log("<color=#FF0000>Please call <color=#FFB800>RealTime.InitStartupTime()</color> in main program first</color>");
                 return false;