+++ Assets/OxGKit/LoggingSystem/Scripts/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## CHANGELOG
 
+## [0.0.2-preview] - 2023-09-09
+- Added LoggingLauncher Editor (can runtime reload setting).
+
 ## [0.0.1-preview] - 2023-08-27
 - Added LoggingSystem.