+++ Assets/OxGKit/Utilities/Scripts/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## CHANGELOG
 
+## [0.0.5-preview] - 2023-09-09
+- Added Requester has clear LRU and ARC methods.
+- Added Cacher has clear methods.
+- Modifier RealTimer create time by itself (don't need to init startup time by RealTime).
+- Modifier RTUpdate create time by itself (don't need to init startup time by RealTime).
+- Optmized Timer.
+
 ## [0.0.4-preview] - 2023-08-24
 - Modifier UMT will auto create an instance use singleton mode.